tests: Move `prepare()` fn to `utils` module
This will allow us to share this function with other tests
