+++ tests/autolib.rs
@@ -1,5 +1,6 @@
 use cargo_manifest::Manifest;
-use tempfile::TempDir;
+
+mod utils;
 
 const BASIC_MANIFEST: &str = r#"
 [package]
@@ -7,32 +8,16 @@ name = "auto-lib"
 version = "0.1.0"
 "#;
 
-fn prepare(manifest: &str, extra_files: Vec<&str>) -> TempDir {
-    let tempdir = tempfile::tempdir().unwrap();
-
-    // Create `Cargo.toml` manifest file
-    std::fs::write(tempdir.path().join("Cargo.toml"), manifest).unwrap();
-
-    // Create extra files
-    for file in extra_files {
-        let path = tempdir.path().join(file);
-        std::fs::create_dir_all(path.parent().unwrap()).unwrap();
-        std::fs::write(path, "").unwrap();
-    }
-
-    tempdir
-}
-
 #[test]
 fn test_bin() {
-    let tempdir = prepare(BASIC_MANIFEST, vec!["src/main.rs"]);
+    let tempdir = utils::prepare(BASIC_MANIFEST, vec!["src/main.rs"]);
     let m = Manifest::from_path(tempdir.path().join("Cargo.toml")).unwrap();
     assert!(m.lib.is_none());
 }
 
 #[test]
 fn test_lib_rs() {
-    let tempdir = prepare(BASIC_MANIFEST, vec!["src/lib.rs"]);
+    let tempdir = utils::prepare(BASIC_MANIFEST, vec!["src/lib.rs"]);
     let m = Manifest::from_path(tempdir.path().join("Cargo.toml")).unwrap();
 
     let lib = m.lib.unwrap();
@@ -52,7 +37,7 @@ fn test_name_override() {
     [lib]
     name = "foo"
     "#;
-    let tempdir = prepare(manifest, vec!["src/lib.rs"]);
+    let tempdir = utils::prepare(manifest, vec!["src/lib.rs"]);
     let m = Manifest::from_path(tempdir.path().join("Cargo.toml")).unwrap();
 
     let lib = m.lib.unwrap();
@@ -72,7 +57,7 @@ fn test_path_override() {
     [lib]
     path = "src/foo.rs"
     "#;
-    let tempdir = prepare(manifest, vec!["src/foo.rs", "src/lib.rs"]);
+    let tempdir = utils::prepare(manifest, vec!["src/foo.rs", "src/lib.rs"]);
     let m = Manifest::from_path(tempdir.path().join("Cargo.toml")).unwrap();
 
     let lib = m.lib.unwrap();
@@ -94,7 +79,7 @@ fn test_other_override() {
     proc-macro = true
     test = false
     "#;
-    let tempdir = prepare(manifest, vec!["src/lib.rs"]);
+    let tempdir = utils::prepare(manifest, vec!["src/lib.rs"]);
     let m = Manifest::from_path(tempdir.path().join("Cargo.toml")).unwrap();
 
     let lib = m.lib.unwrap();
